+++ php-ext/main.cpp
@@ -216,10 +216,10 @@ pmeasure::~pmeasure()
 	using namespace std::chrono;
 	static size_t rcount = 0;
 	auto epd = end_ts.time_since_epoch();
-	fprintf(fh, "%6d %9zu %llu.%03lu: %9lldµs %s\n", getpid(), ++rcount,
-		static_cast<unsigned long long>(duration_cast<seconds>(epd).count()),
-		duration_cast<milliseconds>(epd).count() % std::milli::den,
-		static_cast<unsigned long long>(duration_cast<microseconds>(end_ts - start_ts).count()),
+	fprintf(fh, "%6d %9zu %lld.%03ld: %9lldµs %s\n", getpid(), ++rcount,
+		static_cast<long long>(duration_cast<seconds>(epd).count()),
+		static_cast<long>(duration_cast<milliseconds>(epd).count() % std::milli::den),
+		static_cast<long long>(duration_cast<microseconds>(end_ts - start_ts).count()),
 		what.c_str());
 	fclose(fh);
 }
